LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-focusability-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: Focus order mismatch at step 2/11 expected Element node <button id="within1">button</button> but got Element node <dialog tabindex="0" id="dialog2" open="">
  <button id="...
before button after 1 button after 2 button after 3 button  after 4

Harness Error (FAIL), message = Unhandled rejection: assert_equals: Focus order mismatch at step 2/11 expected Element node <button id="within1">button</button> but got Element node <dialog tabindex="0" id="dialog2" open="">
  <button id="...

TIMEOUT The dialog element itself should not be keyboard focusable. Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: Focus order mismatch at step 2/11 expected Element node <button id="within1">button</button> but got Element node <dialog tabindex="0" id="dialog2" open="">
  <button id="...

TIMEOUT The dialog element itself should not be keyboard focusable. Test timed out
